+++ src/number.h
@@ -6,11 +6,11 @@ scm_object *scm_number_new_integer(const char *num, int radix);
 scm_object *scm_number_new_integer_from_float(const char *num);
 scm_object *scm_number_new_float(const char *num);
 scm_object *scm_number_new_float_from_integer(const char *num, int radix);
-scm_object *INTEGER(int n);
+scm_object *INTEGER(long n);
 scm_object *FLOAT(double n);
 
 char *scm_number_to_string(scm_object *obj, int radix);
-int scm_integer_get_val(scm_object *obj);
+long scm_integer_get_val(scm_object *obj);
 double scm_float_get_val(scm_object *obj);
 
 int scm_number_init(void);
